Prevent error while flushing
It fixes the following error: `Uncaught TypeError: tag.parentNode is null` raising in the `flush()` function when `tag.parentNode` is not defined. This happens when the tag has been already moved / removed as a side-effect. For example, it happens when an application uses [Turbolinks](https://github.com/turbolinks/turbolinks) as page-switcher library, which moves around tags by its own.